requirements.txt
torch==1.10.0
tensorboardX
pyquaternion
nuscenes-devkit
cffi
llvmlite==0.31.0
numba==0.47.0
torchgeometry
numpy==1.19
PyYAML
matplotlib
open3d
opencv-python
Pillow
pyquaternion
scikit-learn
scipy
tqdm